Move CookieParser to Common
