+++ src/System.Net.Primitives/tests/UnitTests/System.Net.Primitives.UnitTests.Tests.csproj
@@ -50,6 +50,9 @@
     <Compile Include="..\..\src\System\Net\Sockets\AddressFamily.cs">
       <Link>ProductionCode\System\Net\Sockets\AddressFamily.cs</Link>
     </Compile>
+    <Compile Include="$(CommonPath)\System\Net\CookieParser.cs">
+      <Link>ProductionCode\System\Net\CookieParser.cs</Link>
+    </Compile>
     <Compile Include="$(CommonPath)\System\Net\SocketAddress.cs">
       <Link>ProductionCode\System\Net\SocketAddress.cs</Link>
     </Compile>
